langchain-example-app/requirements.txt
langchain
python-dotenv
openai
requests